Cleanup books into single flat structure
Currently all of the sections are divided into books which were
used by some static website generator. We no longer require this
structure for the sections. So a flatter strucuture is more appropriate.

This change removes the nested structure for the sections and moves them
into a flatter structure with prettier names.